Final Report/Screening resumes for jobs using nlp techniques/requirements.txt
numpy==1.18.5
textdistance==4.3.0
spacy==2.3.4
streamlit==1.10.0
plotly==5.6.0
sentence_transformers==0.4.1.2
textract==1.6.5
gensim==4.2.0
nltk==3.3
matplotlib==2.2.5
wordcloud==1.4.1
pandas==1.1.5
Pillow==9.2.0
scikit_learn==1.1.2
